requirements.txt
numpy
opencv-python
pillow
cryptography
pytest
click
scipy
matplotlib
